# packages/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Nimiq Simple Faucet
  version: 0.0.1
  description: Public, admin, and MCP surfaces for the Nimiq Simple Faucet
    (testnet). See /SECURITY.md for the threat model.
  license:
    name: MIT
    identifier: MIT
  contact:
    name: Security
    url: https://github.com/onmax/Nimiq-Simple-Faucet/blob/main/SECURITY.md
servers:
  - url: /
    description: This instance
...
components:
  securitySchemes:
    integratorHmac:
      type: apiKey
      in: header
      name: X-Faucet-Api-Key
      description: "Integrator HMAC auth: include `X-Faucet-Api-Key`,
        `X-Faucet-Timestamp`, `X-Faucet-Nonce`, and `X-Faucet-Signature`
        (base64-HMAC-SHA256 over
        `method\\npath\\ntimestamp\\nnonce\\nbodyHash`)."
    adminSession:
      type: apiKey
      in: cookie
      name: faucet_session
      description: Session cookie issued by POST /admin/auth/login. Mutating calls
        also require the `X-Faucet-Csrf` double-submit header matching the
        `faucet_csrf` cookie.
    adminMcpToken:
      type: apiKey
      in: header
      name: X-Faucet-Admin-Token
      description: Required to invoke admin-scoped MCP tools over POST /mcp.
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        code:
          type: string
        message:
          type: string
        issues:
          type: array
          items: {}
      required:
        - error
...
    ClaimListItem:
      type: object
      properties:
        id:
          type: string
        createdAt:
          anyOf:
            - type: string
            - type: number
        address:
          type: string
        status:
          type: string
        decision:
          type:
            - string
            - "null"
        txId:
          type:
            - string
            - "null"
        ip:
          type:
            - string
            - "null"
        integratorId:
          type:
            - string
            - "null"
        abuseScore:
          type:
            - integer
            - "null"
        rejectionReason:
          type:
            - string
            - "null"
      required:
        - id
        - createdAt
        - address
        - status
        - decision
        - txId
        - ip
        - integratorId
        - abuseScore
        - rejectionReason
    ClaimListResponse:
      type: object
      properties:
        total:
          type: integer
        items:
          type: array
          items:
            $ref: "#/components/schemas/ClaimListItem"
      required:
        - total
        - items
...
    IntegratorRecord:
      type: object
      properties:
        id:
          type: string
        createdAt:
          anyOf:
            - type: string
            - type: number
        lastUsedAt:
          anyOf:
            - type: string
            - type: number
            - type: "null"
        revokedAt:
          anyOf:
            - type: string
            - type: number
            - type: "null"
      required:
        - id
        - createdAt
        - lastUsedAt
        - revokedAt
    IntegratorCreateRequest:
      type: object
      properties:
        id:
          type: string
          minLength: 1
          maxLength: 64
          pattern: ^[a-zA-Z0-9_-]+$
      required:
        - id
    IntegratorCredential:
      type: object
      properties:
        id:
          type: string
        apiKey:
          type: string
        hmacSecret:
          type: string
      required:
        - id
        - apiKey
        - hmacSecret
...
    OkResponse:
      type: object
      properties:
        ok:
          type: boolean
          enum:
            - true
      required:
        - ok
...
paths:
...
  /admin/claims:
    get:
      tags:
        - Admin
      summary: List claims
      security:
        - adminSession: []
      parameters:
        - schema:
            type:
              - integer
              - "null"
          required: false
          name: limit
          in: query
        - schema:
            type:
              - integer
              - "null"
          required: false
          name: offset
          in: query
        - schema:
            type: string
          required: false
          name: status
          in: query
        - schema:
            type: string
          required: false
          name: decision
          in: query
        - schema:
            type: string
          required: false
          name: address
          in: query
      responses:
        "200":
          description: List
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClaimListResponse"
...
  /admin/integrators:
    get:
      tags:
        - Admin
      summary: List integrator keys
      security:
        - adminSession: []
      responses:
        "200":
          description: List
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/IntegratorRecord"
                required:
                  - items
    post:
      tags:
        - Admin
      summary: Provision an integrator (one-time credential return)
      security:
        - adminSession: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IntegratorCreateRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IntegratorCredential"
        "409":
          description: Exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /admin/claims/{id}/explain:
    get:
      tags:
        - Admin
      summary: Return structured abuse signals for a claim
      security:
        - adminSession: []
      parameters:
        - schema:
            type: string
          required: true
          name: id
          in: path
      responses:
        "200":
          description: Claim with expanded signals
        "404":
          description: Claim not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/claims/{id}/allow:
    post:
      tags:
        - Admin
      summary: Override a claim to allow (manual admin action)
      security:
        - adminSession: []
      parameters:
        - schema:
            type: string
          required: true
          name: id
          in: path
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"
        "404":
          description: Claim not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/claims/{id}/deny:
    post:
      tags:
        - Admin
      summary: Override a claim to deny (manual admin action)
      security:
        - adminSession: []
      parameters:
        - schema:
            type: string
          required: true
          name: id
          in: path
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                reason:
                  type: string
                  maxLength: 256
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"
        "404":
          description: Claim not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /admin/integrators/{id}/rotate:
    post:
      tags:
        - Admin
      summary: Rotate an integrator API key + HMAC secret
      security:
        - adminSession: []
      parameters:
        - schema:
            type: string
          required: true
          name: id
          in: path
      responses:
        "200":
          description: Rotated credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IntegratorCredential"
        "404":
          description: Integrator not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/integrators/{id}:
    delete:
      tags:
        - Admin
      summary: Revoke an integrator (soft-delete)
      security:
        - adminSession: []
      parameters:
        - schema:
            type: string
          required: true
          name: id
          in: path
      responses:
        "200":
          description: Revoked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"
        "404":
          description: Integrator not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...